refactor(frontend): split pass + admin pages into components, hooks, and shared libs
Both page files were monoliths. Extracted:
- app/lib: types, i18n (STRINGS/LangProvider/useT/LangToggle), format helpers, scan constants
- app/components: shared SmileyMark/BrandSign/LiveClock + izakaya art; pass (RSVPCard, DrinkCard, state screens); admin (login, header, filter bar, attendee row, waitlist, scan popup)
- app/hooks: usePass, useAdminData, useScanner
- page.tsx / admin/page.tsx are now thin composition roots

De-duplicated SmileyMark/BrandSign/LiveClock/drink colours that were copy-pasted across both pages. No behaviour or visual change (verified pass EN/JP, admin list, sticky bar, scan popup).

Also un-ignore frontend/app/lib (was shadowed by the Python lib/ rule).
